se agrego servicio de correo

+++ src/main/java/com/carlos/sistema_ventas/dto/product/ProductRequest.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public record ProductRequest(
         @NotNull
-        Integer categoryId,
+        Long categoryId,
         @NotBlank
         @Size(max = 150)
         String name,
