+++ src/main/java/com/carlos/sistema_ventas/dto/sale/SaleRequest.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotNull;
 
 public record SaleRequest(
         @NotNull
-        Integer quotationId
+        Long quotationId
 ) {
 }
